Refactor codes and add authentication on request header

+++ graphql/resolver/booking.js
@@ -4,28 +4,31 @@ const Booking = require('../../models/bookings')
 // Utils
 const {
   deleteBookindById,
-  singleEventUtils,
   findBookingByIdWithEvent,
   findEventById,
   getAllBooking,
-  userUtils
-} = require('../../utils')
+  transformData
+} = require('../utils')
+
+// constants
+const queryType = require('../constants/queryType')
+const {
+  CANCEL_BOOKING,
+  BOOKING
+} = queryType
 
 const eventResolvers = {
-  bookings: async () => {
+  bookings: async (args, req) => {
+    const {isAuth} = req
+
+    if (!isAuth) {
+      throw new Error('Not Authenticated!')
+    }
+
     try {
       const bookings = await getAllBooking()
       const result = bookings.map(async (booking) => {
-        const {_doc} = booking
-        console.log(_doc)
-
-        return {
-          ..._doc,
-          user: userUtils.bind(this, _doc.user),
-          event: singleEventUtils.bind(this, _doc.event),
-          createdAt: new Date(_doc.createdAt).toISOString(),
-          updatedAt: new Date(_doc.updatedAt).toISOString()
-        }
+        return transformData(booking, BOOKING)
       })
 
       return result
@@ -33,39 +36,45 @@ const eventResolvers = {
       throw err
     }
   },
-  bookEvent: async (args) => {
+  bookEvent: async (args, req) => {
+    const {
+      userId,
+      isAuth
+    } = req
+
+    if (!isAuth) {
+      throw new Error('Not Authenticated!')
+    }
+
     const {eventId} = args
 
     try {
       const fetchEvent = await findEventById(eventId)
       const booking = new Booking({
-        user: '5cf3dcda18aa3035afb2373e',
+        user: userId,
         event: fetchEvent
       })
       const result = await booking.save()
 
-      return {
-        ...result._doc,
-        user: userUtils.bind(this, result._doc.user),
-        event: singleEventUtils.bind(this, result._doc.event),
-        createdAt: new Date(result._doc.createdAt).toISOString(),
-        updatedAt: new Date(result._doc.updatedAt).toISOString()
-      }
+      return transformData(booking, BOOKING)
     } catch (err) {
       throw err
     }
   },
-  cancelBooking: async (args) => {
+  cancelBooking: async (args, req) => {
+    const {isAuth} = req
+
+    if (!isAuth) {
+      throw new Error('Not Authenticated!')
+    }
+
     const {bookingId} = args
     try {
       const booking = await findBookingByIdWithEvent(bookingId)
-      console.log(booking)
-      const event = {
-        ...booking.event._doc,
-        creator: userUtils.bind(this, booking.event._doc.creator)
-      }
+      const event = transformData(booking.event, CANCEL_BOOKING)
 
       await deleteBookindById(bookingId)
+
       return event
     } catch (err) {
       throw err
